Use `StreamExt::split` and allow options to be set on `SocketBuilder`
The `split` method in `comm` does not appear to be `Send` and so fails
when trying to use `tokio::spawn`.  The standard `split` method from
futures works fine though.

Some socket options need to be set before bind/connect, so the
`SocketBuilder` should allow them to be set beforehand.  Some trickery
is used to defer errors until connect/bind are called.