Add set_rcvtimeo and set_sndtimeo
from http://api.zeromq.org/3-0:zmq-setsockopt

```
ZMQ_RCVTIMEO: Maximum time before a recv operation returns with EAGAIN
Sets the timeout for receive operation on the socket. If the value is 0, zmq_recv(3) will return immediately, with a EAGAIN error if there is no message to receive. If the value is -1, it will block until a message is available. For all other values, it will wait for a message for that amount of time before returning with an EAGAIN error.

Option value type	int
Option value unit	milliseconds
Default value	-1 (infinite)
Applicable socket types	all
ZMQ_SNDTIMEO: Maximum time before a send operation returns with EAGAIN
Sets the timeout for send operation on the socket. If the value is 0, zmq_send(3) will return immediately, with a EAGAIN error if the message cannot be sent. If the value is -1, it will block until the message is sent. For all other values, it will try to send the message for that amount of time before returning with an EAGAIN error.

Option value type	int
Option value unit	milliseconds
Default value	-1 (infinite)
Applicable socket types	all
```